### 2020-10-23 #2775 @whedon
[PRE REVIEW]: The nnlib2 library and nnlib2Rcpp R package for implementing neural networks
**Submitting author:** @VNNikolaidis (<a href="http://orcid.org/0000-0003-1471-8788">Vasilis Nikolaidis</a>)
**Repository:** <a href="https://github.com/VNNikolaidis/nnlib2Rcpp" target ="_blank">https://github.com/VNNikolaidis/nnlib2Rcpp</a>
**Version:** v0.1.5
