### 2020-10-23 #2775 @kyleniemeyer
Hi @VNNikolaidis, I'm a bit confused about your repository—it is named "joss", when you should really be submitting the repository for one of the packages (probably the main R package). Is that what the repository contains, and it is just not called "nnlib2Rcpp" for some reason?
