### 2020-10-23 #2775 @VNNikolaidis
Dear Kyle, thank you for your comment. You are right, I misinterpreted the instructions in Submission Requirements (this is my first submission on JOSS). They state "Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default)".

Instead, I created a new repository which is a copy of the actual repository, and added the paper there.

The actual repository (containing the main R package as you correctly mentioned, and also the C++ library it uses) is in https://github.com/VNNikolaidis/nnlib2Rcpp

Since I could easily add the paper there (as a branch this time), and resubmit it, maybe that is a good way to correct the issue.
