### 2020-10-23 #2775 @kyleniemeyer
@VNNikolaidis got it—yes, it would be great if you can add the paper to the actual repository (on a branch is fine). We can update the repository URL here, you do not need to resubmit. @openjournals/dev can you help with this?
