### 2020-10-23 #2775 @VNNikolaidis
That's great, I will add the paper to the actual repository (or a branch) soon and get back to you. I would do it now, but I worry not to mess with the delicate R package structure as it is known to be strict, so I need to investigate this a little just to be sure.
Thank you again!
