+++ paper/paper.bib
@@ -1,0 +1,107 @@
+@misc{keras1,
+  title={Keras},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  howpublished={\url{https://keras.io}}
+}
+
+@incollection{PyTorch1,
+title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+author = {Paszke, Adam and Gross, Sam and others},
+booktitle = {Advances in Neural Information Processing Systems 32},
+editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
+pages = {8024--8035},
+year = {2019},
+publisher = {Curran Associates, Inc.},
+}
+
+@book{SimpsonANS,
+author = {Simpson, Patrick K.},
+title = {Artificial Neural Systems: Foundations, Paradigms, Applications, and Implementations},
+year = {1991},
+isbn = {0071053557},
+publisher = {McGraw-Hill, Inc.},
+address = {USA}
+}
+
+@book{McCord91,
+author = {McCord-Nelson, Marilyn and Illingworth, W. T.},
+title = {A Practical Guide to Neural Nets},
+year = {1991},
+isbn = {0201523760},
+publisher = {Addison-Wesley Longman Publishing Co., Inc.},
+address = {USA}
+}
+
+@book{pao1989adaptive,
+  title={Adaptive Pattern Recognition and Neural Networks},
+  author={Pao, Yohhan},
+  year={1989},
+  publisher={Reading, MA (US); Addison-Wesley Publishing Co., Inc.}
+}
+
+@book{TheoPR,
+author = {Theodoridis, Sergios and Koutroumbas, Konstantinos},
+title = {Pattern Recognition, Fourth Edition},
+year = {2008},
+isbn = {1597492728},
+publisher = {Academic Press, Inc.},
+address = {USA},
+edition = {4th}
+}
+
+
+@article{philippidis1999unsupervised,
+  title={Unsupervised Pattern Recognition Techniques for the Prediction of Composite Failure},
+  author={Philippidis, TP and Nikolaidis, VN and Kolaxis, JG},
+  journal={Journal of Acoustic Emission},
+  volume={17},
+  number={1-2},
+  pages={69--81},
+  year={1999}
+}
+
+@phdthesis{Nikolaidis199Phd,
+    title    = {Non-Destructive Inspection (NDI) Techniques for Composite Materials using Unconventional Pattern Regognition Methods (in Greek)},
+    school   = { University of Patras},
+    author   = {Nikolaidis, Vasilios},
+    year     = {1999},
+url = {https://www.didaktorika.gr/eadd/handle/10442/11158?locale=en}
+}
+
+@article{nikolaidis2013ans,
+  title={ANS-based Preprocessing of Company Performance Indicators},
+  author={Nikolaidis, Vasilios N and Makris, Ilias A and Stavroyiannis, Stavros},
+  journal={Global Business and Economics Review},
+  volume={15},
+  number={1},
+  pages={49--58},
+  year={2013},
+  publisher={Inderscience Publishers Ltd}
+}
+
+@misc{nnlib2RcppCRAN,
+    title = {Package ‘nnlib2Rcpp’},
+    author = {Vasilis Nikolaidis},
+    year = {2020},
+    note = {R package},
+    journal = {​CRAN repository},
+    url = {https://CRAN.R-project.org/package=nnlib2Rcpp},
+  }
+
+@misc{nnlib2RcppGitHub,
+    title = {‘nnlib2Rcpp’},
+    author = {Vasilis Nikolaidis},
+    year = {2020},
[…]
