### 2020-10-24 #2775 @VNNikolaidis
Not a long pause I am afraid. To avoid adding a directory that R-build may complain about<https://blog.r-hub.io/2020/05/20/rbuildignore/>, I followed the easy solution and created a branch (JOSS) with the paper content (no other changes). It can be found in https://github.com/VNNikolaidis/nnlib2Rcpp/tree/JOSS
I also removed the copy (joss repository) as it is no longer needed and may confuse someone.
Please let me know if this is ok. Thanks again!
