### 2020-11-10 #2775 @kakiac
👋  @VNNikolaidis many thanks for your submission! From a quick inspection of this submission I can see that it meets our [submission criteria](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) - and shows great potential of doing Deep Learning and NNs for R geeks! 🤓

It is however shorter than our JOSS typical papers, so whilst I am looking for reviewers, can I please ask you to look at some of the published papers here https://joss.theoj.org/papers/published to get inspiration on how to augment/expand your paper? Thank you 👍
