### 2020-11-10 #2775 @VNNikolaidis
Dear @kakiac<https://github.com/kakiac> I thank you for your interest, feedback and advice.
It is indeed a tool for geeks, I hope there may be some out there who might be interested to use it for experimentation, or -even better- contribute to improve it.
I fully understand your comment about the length of the paper and I agree, when compared to current papers published, it is shorter. I actually tried to be brief, for three reasons:
(a) to closely follow the sample paper (Gala),
(b) to comply with the guidelines mentioning that it should “not focus on new research results”, and that “a full-length paper is not permitted, and software documentation such as API functionality should not be in the paper”.
(c) I personally prefer reading short papers.
I can think of material that could be added, but I wonder if would break (b) above and/or make the paper too long.
Let me investigate this a little. As I am unfamiliar with the process, should material be added to the paper do I simply go ahead and do them on the repository (even now)?
Thank you and best regards,
V.Nikolaidis
