### 2020-11-12 #2775 @schnorr
Hi @kakiac. Thanks for the invitation. I've just checked the PDF generated by dear @whedon  and it is 2-page long. I could do a review of the code and the 2-page manuscript until de end of November. If that is okay for you, let's go ahead.
