### 2020-12-01 #2775 @kakiac
@VNNikolaidis re your question:

> As I am unfamiliar with the process, should material be added to the paper do I simply go ahead and do them on the repository (even now)?

Yes please! Here at JOSS we actually encourage you to continuously improve the paper throughout the reviewing process. Editors and reviewers, we all tend to engage the authors freely in a conversation aimed at improving the paper.
