### 2020-12-01 #2775 @kakiac
re your comments here: https://github.com/openjournals/joss-reviews/issues/2775#issuecomment-724629047, I fully agree with your points. It's not about length per se - but the paper needs to help both the specialist and the non-specialist reader:

- understand the high-level functionality and purpose of the software
- understand the research purpose of the software
- place the current software in context, by reading about other software addressing related needs
- read about (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.

If you could do some initial thinking around it, and make any modifications, and we (the reviewers and I) might make some suggestions on areas for improvement 👍

Thanks in advance,
