### 2020-12-01 #2775 @kakiac
:wave @agisga @billchenxi @MohmedSoudy @paragkulkarni11 would any of you be available to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
