### 2020-12-01 #2876 @whedon
[REVIEW]: The nnlib2 library and nnlib2Rcpp R package for implementing neural networks
**Submitting author:** @VNNikolaidis (<a href="http://orcid.org/0000-0003-1471-8788">Vasilis Nikolaidis</a>)
**Repository:** <a href="https://github.com/VNNikolaidis/nnlib2Rcpp" target ="_blank">https://github.com/VNNikolaidis/nnlib2Rcpp</a>
**Version:** v.0.1.7.a
