### 2020-12-02 #2876 @kakiac
👋 @agisga @billchenxi @MohmedSoudy @paragkulkarni11 would any of you be available to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
