### 2020-12-03 #2876 @kakiac
👋  Dear Lucas (@schnorr) and Mohmed (@MohmedSoudy),

Many thanks for kindly agreeing to review Vasilis' (@VNNikolaidis) submission for the Journal of Open Source Software (JOSS), we’re delighted to have your help. 🥳

There are two checklists at the top of this issue for the reviewers, please use these to track your progress of the review. If you could like more details about the JOSS reviewing process (it's slightly more interesting than for other journals our there!), have a look at the [JOSS reviewer pages](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Many thanks and looking forward to working with you for this submission.
