### 2020-12-04 #2876 @MohmedSoudy
Thank you for inviting me to review this work, regarding the manuscript i
believe that authors should revise the grammar in the paper, also
comparison with available packages on CRAN such as (nnet, deepnet and h2o)
should be mentioned in the statement of need to enrich the manuscript and
elaborate the novelty of author's work. regarding the documentation the
documentation should be user-friendly and handy so i recommend to visualize
the r codes (installation) as r codes. You can find a template in the
following link https://github.com/MohmedSoudy/UniprotR/blob/master/README.md
.

On Thu, Dec 3, 2020 at 5:13 PM Vasilis N. Nikolaidis <
notifications@github.com> wrote:

> As the author, I would like to thank everyone for volunteering their time
> and effort to the review of this paper and software, especially the editor
> @kakiac <https://github.com/kakiac>. This is my first substitution to
> JOSS, and I admit I do not know much about the JOSS reviewing process, so
> any guidance will be appreciated.
>
> As suggested by @kakiac <https://github.com/kakiac> (here
> <https://github.com/openjournals/joss-reviews/issues/2775#issuecomment-736551955>),
> the paper may gain if extended to present further details of the software
> for the specialist and non-specialist reader. I admit I may have written
> this brief paper more as an invitation for possible collaborators than a
> presentation to users.
>
> I have given some thought on how this could be remedied, but hesitate to
> do changes before the reviewers have taken a look at the material, as I am
> afraid it may confuse the reviewing process.
>
> This software is actually a combination of two entities (a C++ library and
> an R package), and also has two target audiences; below, I will explain
> what I could add for each:
>
> (a) this software may be useful to people only want to use the
> ‘nnlib2Rcpp’ R package as is. The R package does contains some predefined,
> ready-to-use (and rather well-known) neural network models. I could easily
> add some demonstration material for those (and simple examples, similar to
> those in the package documentation).
>
> (b) The paper also presents the ‘nnlib’ collection of C++ classes for
> building neural networks. This is definitely a more advanced subject, as
> these classes may be used for experimentation with new, unusual or custom
> models and configurations. Such process also involves creating custom C++
> sub-classes and recompiling the package, steps definitely not easy for the
> non-specialist reader. However this is, in my opinion, an interesting and
> unusual part of the provided functionality and, with the ‘nnlib2RCpp’ R
> package having an interface (the “NN” R module) for manipulating such
> components, this may fit the needs of some more advanced experimenters. I
> can also add some guidance for doing this (I have already written a brief
> blog post on this – it may be found here
> <https://r-posts.com/creating-custom-neural-networks-with-nnlib2rcpp/>)).
> Maybe something similar could be helpful to the article readers.
>
> I would appreciate if the editor and/or the reviewers give me their
> opinions on the above possible changes, and the ‘ok’ to proceed with such
> changes if they find that are needed.
>
