### 2020-12-07 #2876 @kakiac
Dear @MohmedSoudy, many thanks for these comments. Can I ask you to also update the checklist available at the top of this issue (https://github.com/openjournals/joss-reviews/issues/2876) under your username, as you go through with your review? Reviewing for JOSS is slightly different from other journals in that we ask reviewers to also validate/run the code in the submitted repository - see here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, so any specific thoughts on how to improve this are encouraged.

I will be adding your comments issues in the repository to help the authors keep track of what needs to be addressed.

Let me know if you have any questions, or if something is unclear 😄
