### 2020-12-08 #2876 @MohmedSoudy
For me i think much efforts could be done as the author didn't update the
statement of need that compare his package with available packages in CRAN
or Bioconductor and i believe this is a crucial part in the paper to show
the novelty or the contribution of this work.
