+++ README.md
@@ -36,7 +36,7 @@ citation("nnlib2Rcpp")
 
 ---
 
-The ‘nnlib’ library used (and included) in this package is a collection of C++ classes and templates for creating NNs. This library is also available as a standalone project, in GitHub repository (https://github.com/VNNikolaidis/nnlib2). For a (simplified) class-diagram of significant nnlib2 classes and templates see: https://github.com/VNNikolaidis/nnlib2/blob/master/misc/diagram%20of%20main%20classes.png
+The ‘nnlib2’ library used (and included) in this package is a collection of C++ classes and templates for creating NNs. This library is also available as a standalone project, in GitHub repository (https://github.com/VNNikolaidis/nnlib2). For a (simplified) class-diagram of significant nnlib2 classes and templates see: https://github.com/VNNikolaidis/nnlib2/blob/master/misc/diagram%20of%20main%20classes.png
 
 ---
 
