### 2020-12-30 VNNikolaidis/nnlib2Rcpp#2 @schnorr
License files
There are two files with two licenses: `LICENSE` and `LICENSE.md`. The former has a copyright notice pointing to the author, the second contains a MIT license which is an OSI approved one. I suggest to the author:

- [ ] put a link pointing to the https://opensource.org/licenses/MIT text
