### 2020-12-30 VNNikolaidis/nnlib2Rcpp#3 @schnorr
Readme file
Regarding the `README.md` file

- [ ] It may contain better markup to indicate the code lines that should be copy-pasted to install the package.
- [ ] A simple usage example with some R lines that indicate how to use the package with the different neural networks available
