### 2020-12-30 #2876 @schnorr
I completed the General Checks (Repository, License, Contribution and Authorship, Substantial scholarly effort). As the author stated, the R package is a layer to ease the use of the original [nnlib2 package](https://github.com/VNNikolaidis/nnlib2) - also available in the R package source code as ruled by the CRAN submission procedures. I've left two minor issues regarding the [License](https://github.com/VNNikolaidis/nnlib2Rcpp/issues/2) and the [Readme](https://github.com/VNNikolaidis/nnlib2Rcpp/issues/3) files. The "substantial scholarly effort" has been evaluated using the `sloccount` which points to about +5K lines of code (including the `nnlib2` c++ package that has been previously implemented). The way the commit history is organized made it hard to tell the difference between contributions of the `nnlib2` and its R api through Rcpp since there is a [single commit](https://github.com/VNNikolaidis/nnlib2Rcpp/commit/dd49960f95dfca5feafe1d30eeb339253b8a8bca) that pushes perhaps everything at once. If we consider only Rcpp files with `sloccount src/Rcpp*` (assuming those are the real contribution of the R package), we have about 800 sloc. So, the package provides an entry door to the more complete `nnlib2` package, easing its use through the Rcpp interface. For the R community, this is very welcome.
