### 2020-12-30 #2876 @schnorr
Regarding the Functionality perspective (Installation, Functionality, and Performance). **Installation**. I was able to install the package very easily using both the CRAN sources and the development version available in Github using the installation procedures described in the `README.md` file. When executing the very first command using the library - from the blog post - `p <- new("NN")`, I get the message `(Note: NN is under development (beta).)` using both the CRAN and the development version. **Functionality**. I followed the instructions of the [blog post](https://r-posts.com/creating-custom-neural-networks-with-nnlib2rcpp/), as pointed out by the author, to attempt the utilization of the R package. I wasn't able to reproduce the functionality described there (using the development version) because of this:
```
> p$add_connection_set("perceptron")
Adding (empty) set of perceptron connections to topology.
(once topology is complete, use create_connections() to fill it).
Note: Adding connection set failed.
[1] FALSE
Warning message:
In .CppObject$add_connection_set(...) : Unknown connection set type
```
I didn't go much further using those recommendations because of this error. **Performance**. Regarding the performance, I wasn't able to conduct a complete performance evaluation, but [the author is aware](https://github.com/VNNikolaidis/nnlib2Rcpp) of the underlying `nnlib2` library limitations due to its sequential nature (no parallelism whatsoever) as much of the NN requires strong computational capabilities (accelerated HW for instance).
