### 2020-12-30 #2876 @schnorr
Regarding the Documentation (A statement of need, Installation instructions, Example usage, Functionality documentation, Automated tests, Community guidelines). **A statement of need, Installation instructions**. The documentation has two files (as far as I was able to find). The `README.md` contains a brief statement of need and short installation instructions. There is no example usage, functionality documentation, etc in the readme. **Example usage, Functionality Documentation**. The other file is a PDF that serves as documentation for the R package. It is a [PDF file](https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/nnlib2Rcpp_0.1.4.pdf) in the root of the git repo. It follows the strict guidelines for the CRAN, where every function available in the R package is documented. The PDF files contains an extensive documentation with one example usage to guide the user for every class. **Community guidelines** . The author do invite people to contribute to the software package (see the `README.md` file once again) with the list of topics requiring most attention. I wasn't able to find automated tests (continuous integration) or
