### 2020-12-30 #2876 @schnorr
Regarding Software paper (Summary, A statement of need, State of the field, Quality of writing, References). **Summary**. The summary provides a high-level functionality and purpose of the software for a diverse, non-specialist audience. **Statement of need**. The statement of need make it clear that the usage of this library targets students that want to understand the basics of building NN models from scratch, without the aid of more elaborated packages whose simplification is the main target. The text also highlights the R package integration using Rcpp, which is very nice for the target audience to avoid the intrinsic of the C++ language. I believe the author could provide some words on how much functionality of the `nnlib2` library is available through the R package. Is there anything missing? **State of the field**. There is very limited information on how this software package relates to others targeting the same audience. Even if  author provides references to Keras and Pytorch, I wonder whether other NN packages tailored for students do exists. **Quality of writing**. I think the English writing of the 2-page long paper is appropriate. Generally speaking, the paper is very short (only 2 pages) and I have to agree with my reviewer colleague that the software state of the art could be much more explored by listing other software packages with the same goals.
