### 2020-12-30 #2876 @schnorr
Hi @kakiac, I've completed a first review. Since this is my first time as reviewer for JOSS, please let me know if I need to improve the procedure.
