### 2021-01-04 VNNikolaidis/nnlib2Rcpp#2 @VNNikolaidis
Thank you for the comment. The file LICENCE is written as is to be compatible with the CRAN submission rules, while the LICENSE.md is there for GitHub. Of course they both contain the same licensing information, but this has been an known issue (that it is hard to satisfy both in a single file). See for example the discussion in:

https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements

I will look a little more into this issue although my priority is to comply with CRAN (they are quite strict) while also have the files in the GitHub repository as close as possible to those used when creating the final CRAN submission.
