+++ README.md
@@ -39,8 +39,6 @@ For copyright information see LICENSE.md file or DESCRIPTION+LICENSE files (as i
 
 The ‘nnlib2’ library used (and included) in this package is a collection of C++ classes and templates for creating NNs. This library is also available as a standalone project, in GitHub repository (https://github.com/VNNikolaidis/nnlib2). For a (simplified) class-diagram of significant nnlib2 classes and templates see: https://github.com/VNNikolaidis/nnlib2/blob/master/misc/diagram%20of%20main%20classes.png
 
----
-
 For implementing new NN components and models in nnlib2 that can be used in nnlib2Rcpp, see also: 
 
 https://r-posts.com/creating-custom-neural-networks-with-nnlib2rcpp/
