### 2021-01-04 VNNikolaidis/nnlib2Rcpp#3 @VNNikolaidis
Thank you for the comment. The R commands needed to install the package are now emphasized with a little better markup.
As for the second comment (lack of an example), I added in README.md the command needed to open the package documentation in R. In this package documentation each function and class is already documented and accompanied by an example code “snippet”. I hope this is sufficient, let me know if otherwise.
