### 2021-01-04 #2876 @VNNikolaidis
I thank everyone for their time and useful comments, I will start handling them now.
Some of the issues raised at the original repository as part of this review process have already been handled (but left open till I get your feedback).
As I am now about to attempt to handle some of the comments stated above, is it appropriate and/or convenient to the editor and reviewers to place my comments here? If not, how should I proceed?
