### 2021-01-04 #2876 @VNNikolaidis
If I may I would like to add some comments that may aid the reviewer:

>  When executing the very first command using the library - from the blog post - `p <- new("NN")`, I get the message `(Note: NN is under development (beta).)` using both the CRAN and the development version.

Indeed, I have left that message since the NN class is new and being very versatile and I wanted to inform its users that the method it provides may change, while more methods will (and already have been) added. The message does not imply anything beyond that.

> I followed the instructions of the [blog post](https://r-posts.com/creating-custom-neural-networks-with-nnlib2rcpp/), as pointed out by the author, to attempt the utilization of the R package. I wasn't able to reproduce the functionality described there (using the development version) because of this:
>
> ```
> > p$add_connection_set("perceptron")
> Adding (empty) set of perceptron connections to topology.
> (once topology is complete, use create_connections() to fill it).
> Note: Adding connection set failed.
> [1] FALSE
> Warning message:
> In .CppObject$add_connection_set(...) : Unknown connection set type
> ```

This error is expected, and I can understand why it was confusing. The “perceptron” components are included with version v.0.1.5 which is currently at GitHub, so the example should work if the package is installed from there (and not CRAN). The last version currently submitted to CRAN (v.0.1.4) does not have “perceptron”, but please keep in mind that the blog post is ment to explain how to create the missing “perceptron” component (or any other custom component), from scratch, even using this (v.0.1.4) package source code.
