### 2021-01-14 VNNikolaidis/nnlib2Rcpp#3 @VNNikolaidis
I hope the changes made were sufficient, so I am closing the issue
