### 2021-01-14 VNNikolaidis/nnlib2Rcpp#2 @VNNikolaidis
The format used for LICENSE is the one imposed by CRAN. As far as I am aware, CRAN submission rules imply that only a separate  LICENSE.md can be used.
