### 2021-01-15 6bc18ca
With changes according to JOSS reviewers
