+++ paper/paper.bib
@@ -1,8 +1,23 @@
+@inproceedings{abadi2016tensorflow,
+title={Tensorflow: A system for large-scale machine learning},
+author={M. Abadi and P.Barham and J.Chen and Z.Chen and A.Davis and J.Dean and M.Devin and S.Ghemawat and G.Irving and M.Isard and others},
+booktitle={12th $\{$USENIX$\}$ Symposium on Operating Systems Design and Implementation ($\{$OSDI$\}$ 16)},
+pages={265--283},
+year={2016}
+}
+
+@inproceedings{torch,
+title = {Torch7: A Matlab-like Environment for Machine Learning},
+author = {R. Collobert and K. Kavukcuoglu and C. Farabet},
+booktitle = {BigLearn, NIPS Workshop},
+year = {2011}
+} 
+
 @misc{keras1,
-  title={Keras},
-  author={Chollet, Fran\c{c}ois and others},
-  year={2015},
-  howpublished={\url{https://keras.io}}
+title={Keras},
+author={Chollet, Fran\c{c}ois and others},
+year={2015},
+howpublished={\url{https://keras.io}}
 }
 
 @incollection{PyTorch1,
@@ -13,6 +28,123 @@ editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesing
 pages = {8024--8035},
 year = {2019},
 publisher = {Curran Associates, Inc.},
+}
+
+@Manual{h2o_platform,
+    title = {H2O: Scalable Machine Learning Platform},
+    author = {H2O.ai},
+    year = {2020},
+    note = {version 3.30.0.6},
+    url = {https://github.com/h2oai/h2o-3},
+}
+
+@Inbook{Zell1994,
+author="Zell, Andreas
+and Mache, Niels
+and H{\"u}bner, Ralf
+and Mamier, G{\"u}nter
+and Vogt, Michael
+and Schmalzl, Michael
+and Herrmann, Kai-Uwe",
+editor="Skrzypek, Josef",
+title="SNNS (Stuttgart Neural Network Simulator)",
+bookTitle="Neural Network Simulation Environments",
+year="1994",
+publisher="Springer US",
+address="Boston, MA",
+pages="165--186",
+isbn="978-1-4615-2736-7",
+doi="10.1007/978-1-4615-2736-7_9",
+url="https://doi.org/10.1007/978-1-4615-2736-7_9"
+}
+
+@article{FANN,
+author = {Nissen, Steffen},
+year = {2003},
+month = {12},
+pages = {},
+title = {Implementation of a Fast Artificial Neural Network library (FANN)},
+url = {https://github.com/libfann/fann}
+}
+
+@misc{chollet2017kerasR,
+title={R Interface to Keras},
+author={Chollet, Fran\c{c}ois and Allaire, JJ and others},
+year={2017},
+publisher={GitHub},
+howpublished={\url{https://github.com/rstudio/keras}},
+}
+
+@misc{CRANtorch,
+title = {Package ‘torch’},
+author = {D. Falbel and J. Luraschi and others},
+year = {2020},
+note = {R package},
+journal = {​CRAN repository},
+url = {https://CRAN.R-project.org/package=torch},
+}
+
+@misc{CRANrTorch,
+title = {Package ‘rTorch’},
+author = {A. Reyes},
+year = {2020},
+note = {R package},
+journal = {​CRAN repository},
+url = {https://CRAN.R-project.org/package=rTorch},
+}
+
+@Manual{CRANh2o,
+title = {h2o: R Interface for the 'H2O' Scalable Machine Learning Platform},
+author = {E. LeDell and others},
+year = {2020},
+note = {R package},
[…]
