+++ paper/paper.md
@@ -21,6 +21,20 @@ Artificial neural networks are computing models used in various data-driven appl
 
 # Statement of need
 
-The `nnlib2` library is a collection of C++ base classes and templates for building neural network (NN) components and entire models. While a significant number of very capable, flexible, high performance tools for NN are available today, including frameworks s.a. Keras [@keras1], PyTorch [@PyTorch1] etc, the `nnlib2` library may appeal to NN students and experimenters who seek a basis for implementing various such models in C++ and take advantage of the direct control over the implemented model that this approach allows. Furthermore, the library lets new NN definitions be written in code that is comparatively simple and self-explanatory; most of the provided base classes and templates correspond to typical NN components (processing nodes, layers of nodes, connections, sets of connections between layers, entire neural networks etc) and closely follow the conceptual model often found in related bibliography such as [@SimpsonANS], [@McCord91], [@pao1989adaptive], [@TheoPR] and elsewhere. Typical component functionality and features are provided in these base definitions, while diverse, model-specific components are to be implemented in custom sub-classes. Finally, the library presents a high degree of target and compiler independence, with early versions used in various solutions targeting different operating systems, mainly for research purposes, for example in [@philippidis1999unsupervised], [@Nikolaidis199Phd] and [@nikolaidis2013ans]. To take advantage of the R ecosystem, improve its usability and widen its audience, the `nnlib2` library has also been integrated into an R package (a process supported by `Rcpp` [@Rcpp1]). The package, named `nnlib2Rcpp` (available on CRAN [@nnlib2RcppCRAN] and GitHub [@nnlib2RcppGitHub]) also includes additional R-specific classes and methods that allow its users to instantiate the NN components they define using `nnlib2`, combine them in custom NN configurations and models and generally control, monitor and employ them from R. Finally, the package provides a small collection of predefined components and ready-to-use models (for versions of Back Propagation, autoencoder, supervised/unsupervised Learning Vector Quantization and Matrix Associative Memory neural networks).
+A significant number of capable, flexible, high performance tools for neural networks (NNs) are available today, including frameworks s.a. `Tensorflow` [@abadi2016tensorflow] and `Torch` [@torch], and related high level APIs s.a. `Keras` [@keras1] and `PyTorch` [@PyTorch1]. Ready-to-use NN models can also be found on various machine learning platforms s.a. `H2O` [@h2o_platform] or libraries s.a. `SNNS` [@Zell1994] and `FANN` [@FANN]. NNs are also included in a large number of software packages (`WEKA`, `SPSS`, `Matlab`, `NeuroSolutions`, `Noesis`  and many others). The R language, widely used for data analysis, includes package `nnet` [@CRANnnet] in its standard packages, while a significant number of NN-related extension packages are available. These include interfaces and bindings to aforementioned tools (packages s.a. `keras` [@chollet2017kerasR], `torch` [@CRANtorch], `rTorch` [@CRANrTorch], `h2o ` [@CRANh2o], `RSNNS` [@CRANRSNNS]), as well as several other NN-specific packages that provide ready-to-use NN models (s.a. `deepnet` [@CRANdeepnet], `deepNN` [@CRANdeepNN], RcppDL [@CRANRcppDL] etc.).
+
[…]
