+++ DESCRIPTION
@@ -9,6 +9,6 @@ LazyData: true
 LinkingTo: Rcpp
 Imports: Rcpp , methods
 License: MIT + file LICENSE
-Authors@R: person(given = "Vasilis", family = "Nikolaidis", email = "vnikolaidis@us.uop.gr", role = c("aut", "cph", "cre"), comment = c(ORCID = "0000-0003-1471-8788"))
+Authors@R: person(given = "Vasilis", family = "Nikolaidis", email = "v.nikolaidis@uop.gr", role = c("aut", "cph", "cre"), comment = c(ORCID = "0000-0003-1471-8788"))
 URL: https://github.com/VNNikolaidis/nnlib2Rcpp
 Encoding: UTF-8
