### 2021-01-18 #2876 @VNNikolaidis
Dear Editor, @kakiac , and Reviewers @schnorr , @MohmedSoudy, I wish you a Happy 2021.

As the status of my submission is not very clear, some changes were  performed in accordance to your suggestions and comments.

In particular:

(a) a revised version of the paper was uploaded to the repository, which hopefully addresses some of your comments (s.a. the need to specify what differentiates this package from other similar tools, i.e. the **State of the field** issue). The revised version can be found at branch JOSS of the repository https://github.com/VNNikolaidis/nnlib2Rcpp (same location as before).

(b) following the Reviewer @schnorr suggestions, changes were performed to the repository’s README.md file to clarify installation steps with better markdown and better explain how to access documentation and examples (from package build-in help and CRAN auto-generated reference manual). In addition to the above, a document with instructions and examples (file [‘manual.pdf’)](https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/manual.pdf) was added to the repository and is now mentioned in the README.md to better guide new users of the package. This file will become the package ‘vignette’ in future versions of ‘nnlib2Rcpp’.

(c) for the **Performance** issue, no changes were performed, since the software does not make any performance claims (as was correctly remarked by the reviewer @schnorr). Maybe this fact is sufficient to clear the issue.

(d) for the **Automated tests** issue, the reference manual examples in the build-in R package documentation were examined and verified by Reviewers @schnorr. Since there are no other automated tests, maybe this issue should be cleared.

(e) finally, for the **License** issue, unfortunately CRAN rules impose this format for the MIT license in submitted R packages. Therefore, the LICENCE file is written this way to be compatible with the strict CRAN submission rules where the name of the  copyright holders is listed in a separate file. The LICENSE.md is there for GitHub and all other uses and contains the same licensing information (MIT + the name of copyright holder). As far as I am aware, it is not possible to satisfy CRAN restrictions in a single plain license file that contains both the license and the name of the copyright holder. See for example the discussion in:

https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements

Since these are restrictions imposed by the target repository (CRAN), no changes were performed. However, this being an R package for CRAN, and LICENSE.md does contains the entire license, maybe this issue should be cleared.

Please let me know if these changes fulfill any shortcomings previously mentioned by the Reviewers, and what further changes are necessary to fulfill the JOSS submission requirements.
