+++ src/additional_parts.h
@@ -5,6 +5,57 @@
 namespace nnlib2 {
 
 //--------------------------------------------------------------------------------------------
+// example for manual.pdf
+
+class MEX_connection: public connection
+{
+public:
+
+	// model-specific behavior during mapping stage:
+
+	void recall()
+	{
+		destin_pe().receive_input_value(pow( source_pe().output - weight() , 2) );
+	}
+
+	// model-specific behavior during training stage:
+	// in this example, only the current  connection weight (i.e. weight())
+	// and incoming value from the source node (i.e. source_pe().output) are
+	// used in a series of calculations that eventually change the
+	// connection weight (see last line).
+
+	void encode()
+	{
+		double x  = source_pe().output - weight();
+		double s  = .3;
+		double m  = weight();
+		double pi = 2*acos(0.0);
+		double f  = 1/(s*sqrt(2*pi)) * exp(-0.5*pow((x-m)/s,2));
+		double d = (f * x) / 2;
+		weight() = weight() + d;
+	}
+
+};
+
+typedef Connection_Set < MEX_connection > MEX_connection_set;
+
+//--------------------------------------------------------------------------------------------
+// example for manual.pdf
+
+class MEX_pe : public pe
+ {
+ public:
+
+ void recall()
+ 	{
+ 	pe::recall();
+ 	output = sqrt(output);
+ 	}
+ };
+
+typedef Layer < MEX_pe > MEX_layer;
+
+//--------------------------------------------------------------------------------------------
 // example (for RBloggers blog post): Perceptron components
 
 // Perceptron nodes:
@@ -86,10 +137,11 @@ class example_connection_set_2: public Connection_Set<example_connection>
 // given the name, it should return a pointer to a newly created layer
 //------------------------------------------------------------------------------
 
-        layer PTR generate_custom_layer(string name, int size)
+layer PTR generate_custom_layer(string name, int size)
 {
-	if(name == "JustAdd10")          return new JustAdd10_layer (name,size);
-        if(name == "perceptron")        return new perceptron_layer(name,size);
+	if(name == "JustAdd10")         return new JustAdd10_layer (name,size);
+    if(name == "perceptron")        return new perceptron_layer(name,size);
+    if(name == "MEX")				return new MEX_layer(name, size);
 
 	if(name == "example_layer_0")   return new Layer<example_pe> (name, size);
 	if(name == "example_layer_1")   return new example_pe_layer_1(name, size);
@@ -105,7 +157,8 @@ return NULL;
 
 connection_set PTR generate_custom_connection_set(string name)
 {
-                if(name == "perceptron")                return new perceptron_connection_set(name);
+	if(name == "perceptron")                return new perceptron_connection_set(name);
+	if(name == "MEX")						return new MEX_connection_set(name);
 
 	if(name == "example_connection_set_0")  return new Connection_Set<example_connection>(name);
 	if(name == "example_connection_set_1")  return new example_connection_set_1(name);
