+++ src/Rcpp_NN_module.cpp
@@ -170,12 +170,14 @@ public:
 		layer PTR p = generate_layer(name, size, 0);
 		if(p!=NULL)
 			{
-			m_nn.add_layer(p);
-			TEXTOUT << "Topology changed:\n";
-			outline();
-			return true;
+			if(m_nn.add_layer(p))
+				{
+				TEXTOUT << "Topology changed:\n";
+				outline();
+				return true;
+				}
+			delete p;
 			}
-
 		m_nn.change_is_ready_flag(false);
 		TEXTOUT << "Note: Adding layer failed.\n";
 		return false;
@@ -194,10 +196,13 @@ public:
         connection_set PTR p = generate_connection_set(type, 0);
         if(p!=NULL)
          {
-         m_nn.add_connection_set(p);
-         TEXTOUT << "Topology changed:\n";
-         outline();
-         return true;
+         if(m_nn.add_connection_set(p))
+        	{
+        	TEXTOUT << "Topology changed:\n";
+        	outline();
+        	return true;
+        	}
+         delete p;
          }
 
 		m_nn.change_is_ready_flag(false);
@@ -323,6 +328,105 @@ public:
 	{
 		return m_nn.call_component_recall_all(fwd);
 	}
+
+
+	// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+	// Encode multiple input vectors stored in data set
+
+	bool encode_dataset_unsupervised
+						  (NumericMatrix data,
+                           int pos,							// input component position
+            			   bool fwd = true,
+            			   int epochs = 1000
+						  )
+	{
+		if((data.rows()<=0))
+			{
+			error(NN_DATAST_ERR,"Cannot perform unsupervised training, dataset empty");
+			return false;
+			}
+
+		int num_training_cases=data.rows();
+
+		TEXTOUT << "Encoding (unsupervised)...\n";
+
+		for(int i=0;i<epochs;i++)
+			{
+			if(!m_nn.is_ready())
+				{
+				error(NN_DATAST_ERR,"Training failed");
+				return false;
+				}
+
+			for(int r=0;r<num_training_cases;r++)
+				{
+				if(NOT input_at(pos, data( r , _ ) ))
+					{
+					error(NN_INTEGR_ERR,"Training failed");
+					return false;
+					}
+				encode_all(fwd);
+				}
+			if(i%1000==0) checkUserInterrupt();					// (RCpp function to check if user pressed cancel)
+			}
+
+		TEXTOUT << "Finished.\n";
+		return true;
+		}
+
+
+	// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+	// Decode multiple input vectors stored in data set and get output.
+
+	NumericMatrix recall_dataset(NumericMatrix data_in,
+                            	int input_pos,				// input component position
+                            	int output_pos,				// output component position
+                            	bool fwd = true
+                            	)
+	{
+		NumericMatrix data_out;
+
+		if((input_pos<1) OR (input_pos>size()) OR
+           (output_pos<1) OR (output_pos>size()))
+			{
+			error(NN_INTEGR_ERR,"Invalid component position");
+			return data_out;
+			}
+
+		int in_component_size = sizes()[input_pos-1];
+		int out_component_size = sizes()[output_pos-1];
+		int num_cases = data_in.rows();
+
+		if((num_cases<=0))
+			{
+			error(NN_DATAST_ERR,"Cannot recall (decode or map) empty dataset");
+			return data_out;
+			}
+
+		if((in_component_size!=data_in.cols()) OR
+           (out_component_size<=0))
+			{
[…]
