+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: nnlib2Rcpp
 Type: Package
 Title: A Collection of Neural Networks
-Version: 0.1.5
+Version: 0.1.6
 Author: Vasilis Nikolaidis [aut, cph, cre] (<https://orcid.org/0000-0003-1471-8788>)
 Maintainer: Vasilis Nikolaidis <vnikolaidis@us.uop.gr>
 Description: Contains versions of Autoencoder, BP, LVQ, MAM NN and a module to define custom neural networks.
@@ -9,6 +9,6 @@ LazyData: true
 LinkingTo: Rcpp
 Imports: Rcpp , methods
 License: MIT + file LICENSE
-Authors@R: person(given = "Vasilis", family = "Nikolaidis", email = "v.nikolaidis@uop.gr", role = c("aut", "cph", "cre"), comment = c(ORCID = "0000-0003-1471-8788"))
+Authors@R: person(given = "Vasilis", family = "Nikolaidis", email = "v.nikolaidis@us.uop.gr", role = c("aut", "cph", "cre"), comment = c(ORCID = "0000-0003-1471-8788"))
 URL: https://github.com/VNNikolaidis/nnlib2Rcpp
 Encoding: UTF-8
