+++ src/Rcpp_NN_module.cpp
@@ -176,6 +176,7 @@ public:
 				outline();
 				return true;
 				}
+			warning("Deleting orphan (?) layer");
 			delete p;
 			}
 		m_nn.change_is_ready_flag(false);
@@ -202,6 +203,7 @@ public:
         	outline();
         	return true;
         	}
+         warning("Deleting orphan (?) connection set");
          delete p;
          }
 
@@ -336,11 +338,11 @@ public:
 	bool encode_dataset_unsupervised
 						  (NumericMatrix data,
                            int pos,							// input component position
-            			   bool fwd = true,
-            			   int epochs = 1000
+            			   int epochs = 1000,				// training epochs (presentations of all data)
+            			   bool fwd = true					// processing direction (order) for components in NN
 						  )
 	{
-		if((data.rows()<=0))
+		if(data.rows()<=0)
 			{
 			error(NN_DATAST_ERR,"Cannot perform unsupervised training, dataset empty");
 			return false;
@@ -352,7 +354,7 @@ public:
 
 		for(int i=0;i<epochs;i++)
 			{
-			if(!m_nn.is_ready())
+			if(NOT m_nn.is_ready())
 				{
 				error(NN_DATAST_ERR,"Training failed");
 				return false;
@@ -367,13 +369,71 @@ public:
 					}
 				encode_all(fwd);
 				}
-			if(i%1000==0) checkUserInterrupt();					// (RCpp function to check if user pressed cancel)
+			if(i%100==0) checkUserInterrupt();					// (RCpp function to check if user pressed cancel)
 			}
 
 		TEXTOUT << "Finished.\n";
 		return true;
 		}
 
+	// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+	// Encode multiple (i,j) vector pairs stored in two corresponding data sets
+
+	bool encode_datasets_supervised
+	(NumericMatrix i_data,				// data set, each row is a vector i of vector-pair (i,j)
+	 int i_pos,							// position (in topology) of component to receive i.
+	 bool i_to_misc,					// vector i will be sent to 'input' if FALSE, to internal 'misc' register if TRUE.
+	 NumericMatrix j_data,				// data set, each row is the corresponding vector j of vector-pair (i,j)
+	 int j_pos,							// position (in topology) of component to receive j.
+	 bool j_to_misc,					// vector j will be sent to 'input' if FALSE, to internal 'misc' register if TRUE.
+     int epochs = 1000,					// training epochs (presentations of all data)
+     bool fwd = true					// processing direction (order) for components in NN
+	)
+	{
+		if( (i_data.rows()<=0) OR
+            (j_data.rows()<=0) OR
+            (i_data.rows()!=j_data.rows()) )
+		{
+			error(NN_DATAST_ERR,"Cannot perform supervised training, invalid dataset size(s)");
+			return false;
+		}
+
+		int num_training_pairs=i_data.rows();
+
+		TEXTOUT << "Encoding (supervised)...\n";
+
+		for(int e=0;e<epochs;e++)
+		{
+			if(NOT m_nn.is_ready())
+			{
+				error(NN_DATAST_ERR,"Training failed");
+				return false;
+			}
+
+			bool i_data_sent, j_data_sent = false;
+
+			for(int r=0;r<num_training_pairs;r++)
+			{
+				if(i_to_misc)	i_data_sent = set_misc_values_at(i_pos, i_data( r , _ ));
+				else			i_data_sent = input_at(i_pos, i_data( r , _ ));
+
+				if(j_to_misc)	j_data_sent = set_misc_values_at(j_pos, j_data( r , _ ));
[…]
