+++ man/NN-class.Rd
@@ -29,7 +29,7 @@ This R module maintains a generic neural network that can be manipulated using t
   \item New \code{layer}, \code{connection_set}, \code{pe} or \code{connection} definitions must comply (at least loosely) to the \pkg{nnlib2} base class hierarchy and structure and follow the related guidelines. Note: some minimal examples of class and type definitions can be found in the "\code{additional_parts.h}" file itself.
   \item A textual name must be assigned to any new \code{layer} or \code{connection_set}, to be used as parameter in \code{NN} module methods that require a name to create a component. This can be as simple as a single line of code where given the textual name the corresponding component object is created and returned. This code must be added (as appropriate) to either \code{generate_custom_layer()} or \code{generate_custom_connection_set()} functions found in the same "\code{additional_parts.h}" header file. Note: example entries can be found in these functions at the "\code{additional_parts.h}" file.
 }
-Please consider submitting any components you create to be included in future versions of the package.
+More information on expanding the library with new types of NN components (nodes, layers, connections etc) and models, can be found in the package's  \href{https://github.com/VNNikolaidis/nnlib2Rcpp}{repository on Github}). Please consider submitting any useful components you create, to enrich  future versions of the package.
 }
 
 %% ~Make other sections like Warning with \section{Warning }{....} ~
@@ -85,39 +85,38 @@ n$get_output_from(5)
 
 # Example 2: A simple MAM NN
 
-# (2.A) Create the NN and its components:
+# (2.A) Preparation:
+
+# Create data pairs
+
+iris_data    <- as.matrix( scale( iris[1:4] ) )
+iris_species <- matrix(data=-1, nrow=nrow(iris_data), ncol=3)
+for(r in 1:nrow(iris_data))
+ iris_species[r ,as.integer( iris$Species )[r]]=1
+
+# Create the NN and its components:
 
 m <- new( "NN" )
 m$add_layer( "generic" , 4 )
 m$add_layer( "generic" , 3 )
 m$fully_connect_layers_at(1, 2, "MAM", 0, 0)
 
-# (2.B) Use it to store iris species:
+# (2.B) Use the NN to store iris (data,species) pair:
 
-iris_data    <- as.matrix( scale( iris[1:4] ) )
-iris_species <- as.integer( iris$Species )
-for(r in 1:nrow( iris_data ) )
-{
-	x <- iris_data[r,]
-	z <- rep( -1, 3 )
-	z [ iris_species[r] ] <- 1
-	m$input_at( 1, x )
-	m$input_at( 3, z )
-	m$encode_all( TRUE )
-}
+# encode pairs in NN:
 
-# (2.C) Attempt to recall iris species:
+m$encode_datasets_supervised(
+	iris_data,1,FALSE,
+	iris_species,3,FALSE,
+	1,TRUE)
 
-recalled_ids <- NULL;
-for(r in 1:nrow(iris_data))
-{
-	x <- iris_data[r,]
-	m$input_at( 1, x )
-	m$recall_all( TRUE )
-	z <- m$get_output_from( 3 )
-	recalled_ids <- c( recalled_ids, which.max ( z ) )
-}
+# (2.C) Recall iris species from NN:
 
+recalled_data <- m$recall_dataset(iris_data,1,3,TRUE)
+
+# (2.D) Convert recalled data to ids and plot results:
+
+recalled_ids <- apply(recalled_data, 1, which.max)
 plot(iris_data, pch=recalled_ids)
 }
 
@@ -219,12 +218,25 @@ plot(iris_data, pch=recalled_ids)
     }
     }
 
[…]
