+++ README.md
@@ -27,10 +27,6 @@ Once installed, for package help (including documentation and examples for each 
 
     help(package='nnlib2Rcpp')
 
-More instructions on using 'nnlib2' and 'nnlib2Rcpp' can be found in file "manual.pdf" of the repository, i.e. here:
-
-https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/manual.pdf
-
 A reference manual in PDF format (for the last version in CRAN) can be found here:
 
 https://cran.r-project.org/web/packages/nnlib2Rcpp/nnlib2Rcpp.pdf
@@ -45,7 +41,7 @@ For copyright information see LICENSE.md file or DESCRIPTION+LICENSE files (as i
 
 ---
 
-The ‘nnlib2’ library used (and included) in this package is a collection of C++ classes and templates for creating NNs. This library is also available as a standalone project, in GitHub repository (https://github.com/VNNikolaidis/nnlib2). For a (simplified) class-diagram of significant nnlib2 classes and templates see: https://github.com/VNNikolaidis/nnlib2/blob/master/misc/diagram%20of%20main%20classes.png
+The ‘nnlib2’ library used (and included) in this package is a collection of C++ base classes and templates for creating NNs. This library is also available as a standalone project, in GitHub repository (https://github.com/VNNikolaidis/nnlib2). For a (simplified) class-diagram of significant nnlib2 classes and templates see: https://github.com/VNNikolaidis/nnlib2/blob/master/misc/diagram%20of%20main%20classes.png
 
 For implementing new NN components and models in nnlib2 that can be used in nnlib2Rcpp, see also: 
 
@@ -53,7 +49,9 @@ https://r-posts.com/creating-custom-neural-networks-with-nnlib2rcpp/
 
 ( https://wp.me/p8rgs6-sh )
 
-The included Matrix Associative Memory (MAM) related code is also a good example due to the model's simplicity.
+More instructions on using 'nnlib2' and 'nnlib2Rcpp' can also be found in file "manual.pdf" of the repository, i.e. here:
+
+https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/manual.pdf
 
 ---
 
