+++ src/nn.h
@@ -138,6 +138,7 @@ class nn : public component, public data_receiver, public data_provider, public 
  DATA get_weight_at_component(int index, int connection_number);        // returns 0 if not successful
  bool set_weight_at_component(int index, int connection_number, DATA weight);
  bool set_misc_at_component(int index, DATA * data, int dimension);
+ bool set_output_at_component(int index, DATA * data, int dimension);   // currently only works only for layers
  void change_is_ready_flag(bool new_state);                             // avoid using, nn should set m_nn_is_ready flag itself, once its setup is completed
  };
 
