+++ vignettes/manual.bib
@@ -1,0 +1,212 @@
+@misc{chollet2015keras,
+  title={\pkg{Keras}},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  howpublished={\url{https://keras.io}}
+}
+
+@incollection{NEURIPS2019_9015,
+title = {\pkg{PyTorch}: An Imperative Style, High-Performance Deep Learning Library},
+author = {Paszke, Adam and Gross, Sam and others},
+booktitle = {Advances in Neural Information Processing Systems 32},
+editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
+pages = {8024--8035},
+year = {2019},
+publisher = {Curran Associates, Inc.},
+}
+
+@book{10.5555/129269,
+author = {Blum, Adam},
+title = {Neural Networks in C++: An Object-Oriented Framework for Building Connectionist Systems},
+year = {1992},
+isbn = {0471538477},
+publisher = {John Wiley \& Sons, Inc.},
+address = {USA}
+}
+
+@book{SimpsonANS,
+author = {Simpson, Patrick K.},
+title = {Artificial Neural Systems: Foundations, Paradigms, Applications, and Implementations},
+year = {1991},
+isbn = {0071053557},
+publisher = {McGraw-Hill, Inc.},
+address = {USA}
+}
+
+@book{10.5555/103993,
+author = {McCord-Nelson, Marilyn and Illingworth, W. T.},
+title = {A Practical Guide to Neural Nets},
+year = {1991},
+isbn = {0201523760},
+publisher = {Addison-Wesley Longman Publishing Co., Inc.},
+address = {USA}
+}
+
+@book{pao1989adaptive,
+  title={Adaptive Pattern Recognition and Neural Networks},
+  author={Pao, Yohhan},
+  year={1989},
+  publisher={Reading, MA (US); Addison-Wesley Publishing Co., Inc.}
+}
+
+@Manual{Rcpp1,
+  title = {Rcpp: Seamless R and C++ Integration},
+  author = {Dirk Eddelbuettel and Romain Francois and JJ Allaire and Kevin Ushey and Qiang Kou and Nathan Russell and Douglas Bates and John Chambers},
+  year = {2020},
+  note = {R package version 1.0.4},
+  url = {https://CRAN.R-project.org/package=Rcpp},
+}
+
+@Article{Rcpp2,
+    title = {Extending R with C++: A Brief Introduction to \pkg{Rcpp}},
+    author = {Dirk Eddelbuettel and James Joseph Balamuta},
+    journal = {PeerJ Preprints},
+    year = {2017},
+    month = {aug},
+    volume = {5},
+    pages = {e3188v1},
+    issn = {2167-9843},
+    doi = {10.7287/peerj.preprints.3188v1},
+  }
+
+@misc{Rcpp3,
+  title = {Exposing C++ functions and classes with Rcpp modules},
+  author = {Dirk Eddelbuettel and Romain Francois}, 
+  year = {2017},
+  howpublished = {\url{http://dirk.eddelbuettel.com/code/rcpp/Rcpp-modules.pdf}},
+  note = {Accessed: 2019-12-15}
+}
+
+@Manual{Rlang1,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2019},
+    url = {https://www.R-project.org/}
+  }
+
+@book{kohonen2001,
+  title={Self-Organizing Maps (3rd ed.)},
+  author={Kohonen, Teuvo},
+  year={2001},
+  publisher={Springer}
+}
+
+@book{kohonen2012self,
[…]
