+++ CHANGES.md
@@ -41,11 +41,12 @@ Changes to nnlib2Rcpp version 0.1.5 (from 0.1.4)
 ---
 
 Changes to nnlib2Rcpp version 0.1.6 (from 0.1.5)
--	added manual.pdf
 -	added method to module "NN" (recall_dataset) to decode/map/retrieve output for entire input dataset.
 -	added method to module "NN" (encode_dataset_unsupervised) for faster unsupervised training.
 -	added method to module "NN" (encode_datasets_supervised) for faster supervised training.
 -	includes nnlib2 v.0.2.4
 -	added method to module "NN" (get_output_at).
 -	added method to module "NN" (set_output_at).
+-	renamed secondary method of module "NN" (from set_misc_values_at to set_misc_at).
+-	added package vingette
 -	other minor changes

[…]
