### 2021-02-27 #2876 @VNNikolaidis
Dear  @kakiac (Editor),
I took the liberty to remove the **JOSS** branch of the [nnlib2Rcpp GitHub repository](https://github.com/VNNikolaidis/nnlib2Rcpp) which was created solely for the review of my submission. The (revised) paper is now located at the **master** branch, which contains both the software and all related materials.
Furthermore, I am getting ready to release the new version (v.0.1.6) to CRAN with the changes mentioned in the [previous comment](https://github.com/openjournals/joss-reviews/issues/2876#issuecomment-762440336) - based on the reviewer feedback, as well as other changes - see also [CHANGES.md](https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/CHANGES.md)). The package now includes "vignette" documentation, found (in PDF format) as file [manual.pdf](https://github.com/VNNikolaidis/nnlib2Rcpp/blob/master/doc/manual.pdf).
Finally I would appreciate any conclusive result on my paper submission, so (in case it is positive) use it in the package's CITATION information.
