### 2021-04-05 #2876 @arfon
:wave: @VNNikolaidis - apologies for the delay here - it seems like the handling editor (@kakiac) is struggling to find time to edit right now so I will take over from here. Thanks to @MohmedSoudy and @schnorr for all of their work thus far!

My read of this thread is that this submission is close to being accepted.

The one major issue I see that has not been addressed at this point is the "State of the field" requirements. Quoting from @schnorr's review above:

> State of the field. There is very limited information on how this software package relates to others targeting the same audience. Even if author provides references to Keras and Pytorch, I wonder whether other NN packages tailored for students do exists.

@VNNikolaidis - could you please let us know how you have (or plan to) address this?
