+++ paper/paper.md
@@ -17,24 +17,24 @@ bibliography: paper.bib
 
 # Summary
 
-Artificial neural networks are computing models used in various data-driven applications. Such systems typically consist of a large number of processing elements, usually organized in layers, which transfer data to other nodes via weighted connections. An ever-increasing number of different neural network models has been proposed and used. Among the several factors differentiating each model are the network topology, the processing functions in nodes and connections, the training method used to modify internal state of the network components, the processing sequences utilized for presenting and transferring data to, within and from the model during the training and mapping stages etc. The software presented here is a C++ library of classes and templates for implementing neural network components and models and an R package that allows users to instantiate and use such components directly from the R programming language.
+Artificial neural networks (ANN or NN) are computing models used in various data-driven applications. Such systems typically consist of a large number of processing elements (or nodes), usually organized in layers, which exchange data via weighted connections. An ever-increasing number of different neural network models has been proposed and used. Among the several factors differentiating each model are the network topology, the processing and training methods in nodes and connections, the sequences utilized for transferring data to, within and from the model etc. The software presented herein is a C++ library of classes and templates for implementing neural network components and models and an R package that allows users to instantiate and use such components from the R programming language.
 
 # Statement of need
 
-A significant number of capable, flexible, high performance tools for neural networks (NNs) are available today, including frameworks s.a. `Tensorflow` [@abadi2016tensorflow] and `Torch` [@torch], and related high level APIs s.a. `Keras` [@keras1] and `PyTorch` [@PyTorch1]. Ready-to-use NN models can also be found on various machine learning platforms s.a. `H2O` [@h2o_platform] or libraries s.a. `SNNS` [@Zell1994] and `FANN` [@FANN]. NNs are also included in a large number of software packages (`WEKA`, `SPSS`, `Matlab`, `NeuroSolutions`, `Noesis`  and many others). The R language, widely used for data analysis, includes package `nnet` [@CRANnnet] in its standard packages, while a significant number of NN-related extension packages are available. These include interfaces and bindings to aforementioned tools (packages s.a. `keras` [@chollet2017kerasR], `torch` [@CRANtorch], `rTorch` [@CRANrTorch], `h2o ` [@CRANh2o], `RSNNS` [@CRANRSNNS]), as well as several other NN-specific packages that provide ready-to-use NN models (s.a. `deepnet` [@CRANdeepnet], `deepNN` [@CRANdeepNN], RcppDL [@CRANRcppDL] etc.).
+A significant number of capable, flexible, high performance tools for NN are available today, including frameworks s.a. `Tensorflow` [@abadi2016tensorflow] and `Torch` [@torch], and related high level APIs s.a. `Keras` [@keras1] and `PyTorch` [@PyTorch1]. Ready-to-use NN models are also provided by various machine learning platforms s.a. `H2O` [@h2o_platform] or libraries s.a. `SNNS` [@Zell1994] and `FANN` [@FANN]. Ready to use NNs are available in a large number of software packages (`WEKA`, `SPSS`, `Matlab`, `NeuroSolutions`, `Noesis`  and many others). The R language, widely used for data analysis, includes package `nnet` [@CRANnnet] in its standard packages, while a significant number of NN-related extension packages can be found in CRAN and other repositories. These include interfaces and bindings to aforementioned tools (packages s.a. `keras` [@chollet2017kerasR], `torch` [@CRANtorch], `rTorch` [@CRANrTorch], `h2o ` [@CRANh2o], `RSNNS` [@CRANRSNNS]), as well as several other NN-specific packages that provide ready-to-use NN models (s.a. `deepnet` [@CRANdeepnet], `deepNN` [@CRANdeepNN], RcppDL [@CRANRcppDL] etc.).
 
[…]
