+++ paper/paper.md
@@ -23,17 +23,17 @@ Artificial neural networks (ANN or NN) are computing models used in various data
 
 A significant number of capable, flexible, high performance tools for NN are available today, including frameworks s.a. `Tensorflow` [@abadi2016tensorflow] and `Torch` [@torch], and related high level APIs s.a. `Keras` [@keras1] and `PyTorch` [@PyTorch1]. Ready-to-use NN models are also provided by various machine learning platforms s.a. `H2O` [@h2o_platform] or libraries s.a. `SNNS` [@Zell1994] and `FANN` [@FANN]. Ready to use NNs are available in a large number of software packages (`WEKA`, `SPSS`, `Matlab`, `NeuroSolutions`, `Noesis`  and many others). The R language, widely used for data analysis, includes package `nnet` [@CRANnnet] in its standard packages, while a significant number of NN-related extension packages can be found in CRAN and other repositories. These include interfaces and bindings to aforementioned tools (packages s.a. `keras` [@chollet2017kerasR], `torch` [@CRANtorch], `rTorch` [@CRANrTorch], `h2o ` [@CRANh2o], `RSNNS` [@CRANRSNNS]), as well as several other NN-specific packages that provide ready-to-use NN models (s.a. `deepnet` [@CRANdeepnet], `deepNN` [@CRANdeepNN], RcppDL [@CRANRcppDL] etc.).
 
-The `nnlib2` library adds another tool for NN experimentation and implementation. It is a collection of C++  base classes and class-templates for defining NN parts, components and entire models. It contains base classes and functionality for each part of a NN (processing nodes, connections, layers, groups of connections etc) and for the NN models that consist of these parts. The library does not focus (in its current state) on high-performance computing, nor on providing specific implementations of ready-to-use models (although it does contain some ready-to-use NN models). It aims on being a versatile basis for defining different custom NN parts and models with consistent, readable, maintainable and reusable code. Overall the library (and related R package):
+The `nnlib2` library adds another tool for NN experimentation and implementation. It is a collection of C++  base classes and class-templates for defining NN parts, components and entire models. It contains base classes (and related functionality) for each part of a NN (processing nodes, connections, layers, groups of connections etc) and for NN models that contain such parts. The library does not focus (in its current state) on high-performance computing but on being a versatile basis on which new NN parts and models can be defined with consistent, readable, maintainable and reusable code. Although the software does contain some ready-to-use NN models (currently versions of Back-propagation, autoencoder, Learning Vector Quantization and Matrix Associative Memory) with their related reusable NN parts, its main purpose is to aid the creation of new custom NN parts and the experimentation with them in arbitrary NN model configurations. Overall the library (and related R package):
 
-1. allows for implementation of NN parts and models from reusable components that follow a common interface thus can easily be modified or combined with each other, making them suitable for possible educational or experimentation purposes.
-2. provides a basis for implementation of diverse models, including classic NNs.
+1. allows the implementation of NN parts and models from reusable components that follow a common interface thus can easily be modified, combined with each other etc, making them suitable for educational or experimentation purposes.
+2. provides a basis for implementation of different NN parts and models, including classic NNs.
 3. is versatile enough for experimentation with new, custom components, configurations and models.
 4. has no dependencies with other software; to build or employ the `nnlib2` parts and models only a standard C++ compiler is required, and the produced models are standalone, lightweight, highly portable and can be embedded in any C++ application.
[…]
