### 2021-04-06 #2876 @VNNikolaidis
@arfon - Thank you for your reply and your involvement. The editor’s ( @kakiac ) situation is more than understood, especially under these strange circumstances. I appreciate her time and involvement so far.

Since the initial submission (which was commented by the reviewers) several changes have been done both to the paper and the software, some of which where to remedy issues mentioned by the reviewers. These changes are described in previous posts above ([here](https://github.com/openjournals/joss-reviews/issues/2876#issuecomment-762440336) and below).

In particular, to address the State of Field, I expanded the related “Statement of Need” section to provide more information on what this software is about. With the abundance of neural network tools available an exhaustive comparison is next-to-impossible, and rarely is any tool completely unique, but I hope these changes better clarify what differentiates this package and may make it fit for some user's (and/or possibly contributor's) purposes. However I do not now if these changes suffice, so I would appreciate the feedback.
