### 2021-04-07 #2876 @arfon
> In particular, to address the State of Field, I expanded the related “Statement of Need” section to provide more information on what this software is about. With the abundance of neural network tools available an exhaustive comparison is next-to-impossible, and rarely is any tool completely unique, but I hope these changes better clarify what differentiates this package and may make it fit for some user's (and/or possibly contributor's) purposes. However I do not now if these changes suffice, so I would appreciate the feedback.

Thank you for the update @VNNikolaidis. @schnorr and @MohmedSoudy - could I trouble you both to come and take another look at this submission to decide if you are ready to recommend this submission be accepted?
