### 2021-05-05 afa6a5e
Changes for v.0.1.7
