+++ DESCRIPTION
@@ -1,11 +1,10 @@
 Package: nnlib2Rcpp
 Type: Package
 Title: A Collection of Neural Networks
-Version: 0.1.6
+Version: 0.1.7
 Author: Vasilis Nikolaidis [aut, cph, cre] (<https://orcid.org/0000-0003-1471-8788>)
 Maintainer: Vasilis Nikolaidis <vnikolaidis@us.uop.gr>
 Description: Contains versions of Autoencoder, BP, LVQ, MAM NN and a module to define custom neural networks.
-LazyData: true
 LinkingTo: Rcpp
 Imports: Rcpp , methods
 License: MIT + file LICENSE
