### 2021-05-20 #2876 @MohmedSoudy
I think the paper now is ready to be published after applying the required
modifications.

On Mon, May 10, 2021 at 9:33 AM Mohmed Soudy ***@***.***>
wrote:

> Of course. I will check it.
>
> On Sun, May 9, 2021 at 9:01 PM Arfon Smith ***@***.***>
> wrote:
>
>> @schnorr <https://github.com/schnorr> and @MohmedSoudy
>> <https://github.com/MohmedSoudy> - could I trouble you both to come and
>> take another look at this submission to decide if you are ready to
>> recommend this submission be accepted?
>>
>> Friendly reminder to please check on this submission again sometime soon
>> please @schnorr <https://github.com/schnorr> and @MohmedSoudy
>> <https://github.com/MohmedSoudy>.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/2876#issuecomment-835864447>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AG2XKP6ACKQTZT3UIJ5XVP3TM3LYBANCNFSM4UJEKQKQ>
>> .
>>
>
>
> --
> *Regards,*
> *Mohmed Soudy Mohmed*
>
> *Research Associate*
>
> *Proteomics & Metabolomics Unit*
> *Basic Research*
> *57357 Cancer hospital for children*
>

--
*Regards,*
*Mohmed Soudy Mohmed*

*Research Associate*

*Proteomics & Metabolomics Unit*
*Basic Research*
*57357 Cancer hospital for children*
