### 2021-05-21 #2876 @schnorr
So, I've reviewed my notes above against the latest version of the paper and software. The problems of using the software have been solved because now the CRAN version is aligned with the version expected by the blogpost mentioned previously. The concerns of citing related work have been fixed and the differences against the base C++ implementation are now clear. So, yeah, I think we are ready for publication.
