### 2021-05-22 #2876 @arfon
@VNNikolaidis – forgive me if we've checked these already, but are the DOIs that Whedon is suggesting above correct for your references? If any of them are, please add the DOIs to your BibTeX file, e.g., `doi=10.1163/2214-8647_bnp_e612900`
