+++ paper/paper.bib
@@ -208,6 +208,7 @@ url = {https://www.didaktorika.gr/eadd/handle/10442/11158?locale=en}
   number={1},
   pages={49--58},
   year={2013},
+  doi={10.1504/gber.2013.050667},
   publisher={Inderscience Publishers Ltd}
 }
 
