### 2021-05-22 #2876 @VNNikolaidis
@arfon - on the contrary, thanks for notifying me. Of the 3 DOI suggestions, one is valid (it actually refers to a paper of my own).  The other two DOIs suggested by Whedon are not, they are not the documents referred to by the paper.
I updated the BibTex file on Github but to update Zenodo I took the step to make a new release. This results in a new Zenodo DOI which is DOI: 10.5281/zenodo.4781510
i.e. [![DOI](https://zenodo.org/badge/247648982.svg)](https://zenodo.org/badge/latestdoi/247648982)
Unfortunately I had to use a pseudo-version number (v.0.1.7.a) as I try to keep GitHub release versions same as those submitted to CRAN.
I hope this was the proper way to go, please excuse my ignorance, I am not familiar with the JOSS submission process, this being my first one. Thank you again.
