### 2021-05-23 #2876 @arfon
@VNNikolaidis - a couple of minor additional things. Please merge this PR which makes a few formatting fixes to your paper: https://github.com/VNNikolaidis/nnlib2Rcpp/pull/12

Also, I think you should probably cite the `R` language. You can read about how to do that here: https://intro2r.com/citing-r.html

Finally, don't worry about making a new release for these changes to the paper.
