+++ paper/paper.bib
@@ -75,6 +75,15 @@ publisher={GitHub},
 howpublished={\url{https://github.com/rstudio/keras}},
 }
 
+@Manual{RLanguage,
+title = {R: A Language and Environment for Statistical Computing},
+author = {{R Core Team}},
+organization = {R Foundation for Statistical Computing},
+address = {Vienna, Austria},
+year = {2020},
+url = {https://www.R-project.org/},
+}
+
 @misc{CRANtorch,
 title = {Package ‘torch’},
 author = {D. Falbel and J. Luraschi and others},
