### 2021-05-23 #2876 @VNNikolaidis
I merged the formatting changes (btw, thank you for the effort, time and instructions), added the reference to R language (a valid point) and checked it on the Whedon paper preview service.

About not doing a new release: I realized a release was probably not necessary (since the paper was probably compiled from GitHub) right after I rushed to do yesterday’s release (v.0.1.7.a) only to update Zenodo with the small changes to the paper. But had done it already. No harm done (except for the extra '.a' appended to the version number – the software version is still v.0.1.7). If Whedon does not object, maybe the original Zenodo DOI ( 10.5281/zenodo.4780958 ) is actually a bit more accurate to be used in the paper. If not, it is fine as is, they both refer to the exact same software.
